+++ RecipeApp.API/Endpoints/RecipeEndpoints.cs
@@ -149,6 +149,9 @@ namespace RecipeApp.API.Endpoints
             {
                 var target = await repository.GetById(id);
 
+                if (target == null)
+                    return TypedResults.NotFound();
+
                 if (await repository.Delete(id) != null)
                     return TypedResults.Ok(target);
                 return TypedResults.NotFound();
@@ -406,6 +409,9 @@ namespace RecipeApp.API.Endpoints
                 var recipeIngredients = await repository.GetQueryable(r => r.RecipeId == id).ToListAsync();
 
                 var target = await repository.GetById(id);
+
+                if (target == null)
+                    return TypedResults.NotFound();
 
                 if (await repository.Delete(id) != null)
                     return TypedResults.Ok(target);
