+++ RecipeApp.API/Endpoints/UserEndpoints.cs
@@ -117,6 +117,9 @@ namespace RecipeApp.API.Endpoints
             {
                 var target = await repository.GetById(id);
 
+                if (target == null)
+                    return TypedResults.NotFound();
+
                 if (await repository.Delete(id) != null)
                     return TypedResults.Ok(target);
                 return TypedResults.NotFound();
@@ -191,6 +194,9 @@ namespace RecipeApp.API.Endpoints
             try
             {
                 var target = await repository.GetById(id);
+
+                if (target == null)
+                    return TypedResults.NotFound();
 
                 if (await repository.Delete(id) != null)
                     return TypedResults.Ok(target);
